Remove forward slashes in format lists ...
Although a forward slash has the effect of causing a newline to
be printed when outputting to STDOUT/STDERR, it actually signifies
"the end of a record".

See https://stackoverflow.com/questions/31360099/unclear-cause-of-too-many-records-in-i-o-statement-error

Since newlines seem to be added automatically when logging using the
app package, removing the forward slashes solves the problem and
each message still appears on its own line.