Fix uninitialized warning in grepdep for variable on_continue
The while(on_continue) and the if above it are meant to handle the edge
case where \n is the first character of input.  It seems like this is to
handle empty input which would be a single newline.  Therefore
initializing it to zero will make it so that we will not enter the while
loop only if the first character is \n.  This fits with the intended use
of the program.